refactor(19339): Refatoração dos testes de serviços da prestação de contas
