+++ sme_ptrf_apps/core/tests/tests_services/tests_prestacao_de_contas_services/test_salvar_prestacao_de_contas_service.py
@@ -1,8 +1,7 @@
 import pytest
 
-from ...models.prestacao_conta import STATUS_ABERTO
-from ...models import Observacao
-from ...services import salvar_prestacao_de_contas
+from ....models.prestacao_conta import STATUS_ABERTO
+from ....services import salvar_prestacao_de_contas
 
 pytestmark = pytest.mark.django_db
